RENAME / db/dbBase to db/index
