+++ .env.example
@@ -1,0 +1,8 @@
+clientID=
+clientSecret=
+host=
+database=
+password=
+username=
+WEATHERBIT_TOKEN=
+IPSTACK_TOKEN=
