+++ Server/passport.js
@@ -1,6 +1,6 @@
 const passport = require('passport');
 const GoogleStrategy = require('passport-google-oauth').OAuth2Strategy;
-const addJournals = require('./db/dbBase.js');
+const addJournals = require('./db/index.js');
 require('dotenv').config();
 
 // Use the GoogleStrategy within Passport.
